requirements-in.txt
numpy<2.0
scipy
pillow
opencv-python>=4.8.0
imageio

# Device connection
adbutils
uiautomator2
uiautomator2cache
wrapt
retrying
jellyfish
pyyaml
inflection
pywebio
starlette
anyio
uvicorn[standard]
aiofiles
pypresence
alas-webapp
rich
zerorpc
pyzmq
onepush
pydantic
psutil
chardet
mxnet-alas
matplotlib
pycryptodome
typing_extensions>=4.8.0
pydantic>=2.5.0
onnxruntime; sys_platform != "win32"
onnxruntime-directml; sys_platform == "win32"
rapidocr-onnxruntime